Config structure tweaks
